requirements.txt
streamlit
ultralytics
opencv-python
torch
numpy
Pillow
